[FIX] Preserve voice message if recording is interrupted on IOS
Fixes #3002

When an event interrupts an ongoing voice recording in IOS, the OS stops the recording. This was not handled properly before: The message was immediately completely lost and it was not possible to make a new voice message before force-quitting and restarting the app (because the Expo recorder was left in loaded state).

This PR stops the recording in this case, indicating it was stopped. The user can send or discard the voice message up to that point at any time afterwards.

I also changed the GUI to make it more clear, I'll put this into the comments.